Working simulation code
git-svn-id: svn://localhost/tmp/testgc/trunk@146 230b2fa6-915b-415b-b5c9-3ad8ff460498

+++ simulator/src/com/google/code/laserswarm/conf/Constellation.java
@@ -29,7 +29,7 @@ public class Constellation {
 		receivers.add(new Satellite());
 	}
 
-	public Constellation(double P, int pulseFrequency, Satellite emit, List<Satellite> recv) {
+	public Constellation(double P, double pulseFrequency, Satellite emit, List<Satellite> recv) {
 		power = P;
 		emitter = emit;
 		receivers = recv;
